Initial GeoDistance.java (Java source file) upload
Module: GeoDistance.java | Class Lib
Description: Methods to calculate the great circle (orthodromic) distance between two geo-points on Earth
Version: 17.1.001
Copyright:  2011-2025 Alexander Bell
DISCLAIMER: This Module is provided on AS IS basis without any warranty. The user assumes the entire risk as to the accuracy and the use of this module. In no event shall the author be liable for any damages arising out of the use of or inability to use this module.
TERMS OF USE: This module is copyrighted. Please keep the Copyright notice intact.